requirements.txt
pyautogen>=0.2.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
atlassian-python-api>=3.41.0
requests>=2.31.0
python-json-logger>=2.0.7
termcolor>=2.3.0
typing-extensions>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
regex>=2023.0.0
tqdm>=4.65.0
scipy>=1.15.2
